Update getType() override in File classes

+++ src/main/java/pl/jkuznik/computer/software/file/AbstractFile.java
@@ -2,9 +2,9 @@ package pl.jkuznik.computer.software.file;
 
 public abstract class AbstractFile implements File {
     protected String name;
-    protected int size;
+    protected long size;
 
-    protected AbstractFile(String name, int size) {
+    protected AbstractFile(String name, long size) {
         this.name = name;
         this.size = size;
     }
@@ -15,7 +15,7 @@ public abstract class AbstractFile implements File {
     }
 
     @Override
-    public int getSize() {
+    public long getSize() {
         return this.size;
     }
